adb: Fix handling of writes to register 3
We don't allow an address change without the special handler IDs
otherwise MacOS 9 with a PMU will cause us to incorrectly change
the mouse address to a collision course with the keyboard when it
sends a weird 3 bytes write to that register during boot.

We also limit the handler changes to supported handlers.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>